lets say its smaller

+++ header/map.h
@@ -9,21 +9,21 @@ using namespace std;
 
 bool possibleParts[12][3][3] = {{{0,1,0},{1,1,0},{0,0,0}},  {{0,1,0},{0,1,1},{0,0,0}},  {{0,0,0},{1,1,0},{0,1,0}},  {{0,0,0},{0,1,1},{0,1,0}},  {{0,1,0},{0,1,0},{0,1,0}},  {{0,0,0},{1,1,1},{0,0,0}},  {{0,1,0},{1,1,0},{0,1,0}},  {{0,1,0},{0,1,1},{0,1,0}},  {{0,1,0},{1,1,1},{0,0,0}},  {{0,0,0},{1,1,1},{0,1,0}},  {{0,1,0},{1,1,1},{0,1,0}}, {{0,0,0},{0,0,0},{0,0,0}}};
 
-bool roadIsLeft(array<array<int, WIDTH>, HIGH> field){
-  for(int y = 0; y < HIGH; y++){
-    for(int x = 0; x < WIDTH; x++){
+bool roadIsLeft(array<array<uint8_t, WIDTH>, HIGH> field){
+  for(uint8_t y = 0; y < HIGH; y++){
+    for(uint8_t x = 0; x < WIDTH; x++){
       if(field[y][x] == ROAD) return true;
     }
   }
   return false;
 }
 
-bool FieldIsViable(array<array<int, WIDTH>, HIGH> field){
+bool FieldIsViable(array<array<uint8_t, WIDTH>, HIGH> field){
 
-  for(int y = 0; y < HIGH; y++){
-    for(int x = 0; x < WIDTH; x++){
+  for(uint8_t y = 0; y < HIGH; y++){
+    for(uint8_t x = 0; x < WIDTH; x++){
       if(field[y][x] == ROAD){
-        int neighbourRoadCounter = 0;
+        uint8_t neighbourRoadCounter = 0;
         if(field[y+1][x] == 2) neighbourRoadCounter++;
         if(field[y-1][x] == 2) neighbourRoadCounter++;
         if(field[y][x+1] == 2) neighbourRoadCounter++;
@@ -34,18 +34,18 @@ bool FieldIsViable(array<array<int, WIDTH>, HIGH> field){
   }
 
   for(;;){
-    int a = rand() % HIGH, b = rand() % WIDTH;
+    uint8_t a = rand() % HIGH, b = rand() % WIDTH;
     if(field[a][b] == ROAD){
       field[a][b] = ROAD+1;
       break;
     }
   }
 
-  array<array<int, WIDTH>, HIGH> oldfield = field;
+  array<array<uint8_t, WIDTH>, HIGH> oldfield = field;
   for(;;){
     oldfield = field;
-    for(int y = 0; y < HIGH; y++){
-      for(int x = 0; x < WIDTH; x++){
+    for(uint8_t y = 0; y < HIGH; y++){
+      for(uint8_t x = 0; x < WIDTH; x++){
         if(field[y][x] == ROAD+1){
           if(field[y-1][x] == ROAD) field[y-1][x] = ROAD+1;
           if(field[y+1][x] == ROAD) field[y+1][x] = ROAD+1;
@@ -61,7 +61,7 @@ bool FieldIsViable(array<array<int, WIDTH>, HIGH> field){
 }
 
 
-bool partIsPossible(int partNumber, int fieldTop, int fieldLeft, int fieldBottom, int fieldRight){
+bool partIsPossible(uint8_t partNumber, uint8_t fieldTop, uint8_t fieldLeft, uint8_t fieldBottom, uint8_t fieldRight){
   if(fieldTop != 12){
     if(!(possibleParts[fieldTop][2][1] == possibleParts[partNumber][0][1])) return false;
   }
@@ -77,11 +77,11 @@ bool partIsPossible(int partNumber, int fieldTop, int fieldLeft, int fieldBottom
   return true;
 }
 
-array<array<int, WIDTH>, HIGH> theFinalField(array<array<int, WIDTH/3+2>, HIGH/3+2> fieldParts){
-  array<array<int, WIDTH>, HIGH> field;
+array<array<uint8_t, WIDTH>, HIGH> theFinalField(array<array<uint8_t, WIDTH/3+2>, HIGH/3+2> fieldParts){
+  array<array<uint8_t, WIDTH>, HIGH> field;
 
-  for(int i = 0; i < HIGH; i++){
-    for(int a = 0; a < WIDTH; a++){
+  for(uint8_t i = 0; i < HIGH; i++){
+    for(uint8_t a = 0; a < WIDTH; a++){
       field[i][a] = possibleParts[fieldParts[i/3+1][a/3+1]][i % 3][a % 3]+1;
     }
   }
@@ -89,38 +89,38 @@ array<array<int, WIDTH>, HIGH> theFinalField(array<array<int, WIDTH/3+2>, HIGH/3
   return field;
 }
 
-bool MatchingPartExists(array<array<int, WIDTH/3+2>, HIGH/3+2> fieldParts, int y, int x){
-  for(int i = 0; i < 11; i++){
+bool MatchingPartExists(array<array<uint8_t, WIDTH/3+2>, HIGH/3+2> fieldParts, uint8_t y, uint8_t x){
+  for(uint8_t i = 0; i < 11; i++){
     if(partIsPossible(i, fieldParts[y-1][x], fieldParts[y][x-1], fieldParts[y+1][x], fieldParts[y][x+1])) return true;
   }
   return false;
 }
 
-array<array<int, WIDTH>, HIGH> generateField(){
+array<array<uint8_t, WIDTH>, HIGH> generateField(){
   for(;;){
-    array<array<int, WIDTH/3+2>, HIGH/3+2> fieldParts;
+    array<array<uint8_t, WIDTH/3+2>, HIGH/3+2> fieldParts;
     srand(time(NULL));
 
     //clear field
-    for(int i = 0; i < HIGH/3+2; i++){
-      for(int a = 0; a < WIDTH/3+2; a++){
+    for(uint8_t i = 0; i < HIGH/3+2; i++){
+      for(uint8_t a = 0; a < WIDTH/3+2; a++){
         fieldParts[i][a] = 12;
       }
     }
 
     //rand for fieldParts
-    for(int i = 0; i < HIGH/3+2; i++){
+    for(uint8_t i = 0; i < HIGH/3+2; i++){
       fieldParts[i][0] = 11;
       fieldParts[i][WIDTH/3+1] = 11;
     }
-    for(int i = 0; i < WIDTH/3+2; i++){
+    for(uint8_t i = 0; i < WIDTH/3+2; i++){
       fieldParts[0][i] = 11;
       fieldParts[HIGH/3+1][i] = 11;
     }
 
-    for(int y = 1; y < HIGH/3+1; y++){
-      for(int x = 1; x < WIDTH/3+1; x++){
-        int partNumber;
+    for(uint8_t y = 1; y < HIGH/3+1; y++){
+      for(uint8_t x = 1; x < WIDTH/3+1; x++){
+        uint8_t partNumber;
         if(MatchingPartExists(fieldParts, y, x)){
           for(;;){
             partNumber = rand() % 11;
@@ -133,7 +133,7 @@ array<array<int, WIDTH>, HIGH> generateField(){
       }
     }
 
-    array<array<int, WIDTH>, HIGH> field = theFinalField(fieldParts);
+    array<array<uint8_t, WIDTH>, HIGH> field = theFinalField(fieldParts);
     if(FieldIsViable(field)) return field;
   }
 }
