+++ pacman.cpp
@@ -13,7 +13,7 @@ void getColors(){
   fieldTextures[WALL] = "\033[48;5;2m  ";
 }
 
-void draw(array<array<int, WIDTH>, HIGH> field){
+void draw(array<array<uint8_t, WIDTH>, HIGH> field){
   for(int i = 0; i < HIGH; i++){
     for(int a = 0; a < WIDTH; a++){
       cout << fieldTextures[field[i][a]] << "\033[0m";
@@ -25,7 +25,7 @@ void draw(array<array<int, WIDTH>, HIGH> field){
 int main(int argc, char const *argv[]) {
   //pacman tbd
   getColors();
-  array<array<int, WIDTH>, HIGH> field = generateField();
+  array<array<uint8_t, WIDTH>, HIGH> field = generateField();
   draw(field);
   return 0;
 }
